service/gsp: Move service/gsp_lcd.* into service/gsp/ folder
